perf(shared/useScrollLock): 退场时延迟解锁滚动锁定
