+++ packages/shared/src/hooks/useScrollLock.ts
@@ -4,14 +4,19 @@ import { noScroll } from '@tool-pack/dom';
 export function useScrollLock(
   visible: boolean | void,
   container?: HTMLElement | (() => HTMLElement | undefined),
-  enable = true,
+  options: { enabled?: boolean; delay?: number } = {},
 ) {
+  const { enabled = true, delay = 500 } = options;
   useEffect(() => {
-    if (!enable || !visible) return;
+    if (!enabled || !visible) return;
     const el = typeof container === 'function' ? container() : container;
-    return noScroll(
+    const unlock = noScroll(
       el,
       [document.body, document.documentElement].includes(el!),
     );
-  }, [enable, container, visible]);
+    return () => {
+      // 需要等窗体退场动画结束后才能取消滚动锁定
+      setTimeout(unlock, delay);
+    };
+  }, [enabled, container, visible]);
 }
